[add]動的かつ一意なURLの設定成功
