+++ services/soso-web/src/app/calendarList/page.tsx
@@ -132,9 +132,22 @@ export default function MyPage() {
     alert('ロゴがクリックされました');
   };
 
-  const onCalendarClick = (id: string) => {
-    router.push('/dashboard/' + id);
-  };
+  // ...existing code...
+
+const onCalendarClick = (id: string) => {
+  // ★ 修正: [id]/[slug] 構造に対応
+  // calendar.name をエンコードして slug として使用
+  const calendar = calendars.find(cal => cal.id.toString() === id);
+  if (calendar) {
+    const encodedName = encodeURIComponent(calendar.name);
+    router.push(`/dashboard/${id}/${encodedName}`);
+  } else {
+    // フォールバック: calendar が見つからない場合はデフォルト名を使用
+    router.push(`/dashboard/${id}/calendar`);
+  }
+};
+
+// ...existing code...
 
   const handleAddCalendar = () => {
     setIsTeamAddModalOpen(true);
